fix(dashboard): clean api in dashboard
