+++ dashboard/src/components/TableCustomFields.js
@@ -4,8 +4,8 @@ import { toastr } from 'react-redux-toastr';
 import { Col, FormGroup, Input, Label, Modal, ModalBody, ModalHeader, Row } from 'reactstrap';
 import styled from 'styled-components';
 import { useAuth } from '../recoil/auth';
-import useApi from '../services/api-interface-with-dashboard';
 import { typeOptions } from '../utils';
+import useApi from '../services/api';
 import ButtonCustom from './ButtonCustom';
 import SelectCustom from './SelectCustom';
 import Table from './table';
