+++ dashboard/src/recoil/actions.js
@@ -1,6 +1,6 @@
 import { atom, useRecoilState } from 'recoil';
 import { useAuth } from '../recoil/auth';
-import useApi from '../services/api-interface-with-dashboard';
+import useApi from '../services/api';
 import { getData, useStorage } from '../services/dataManagement';
 import { capture } from '../services/sentry';
 import { useComments } from './comments';
