+++ dashboard/src/recoil/persons.js
@@ -1,7 +1,7 @@
 /* eslint-disable react-hooks/exhaustive-deps */
 import { atom, selector, useRecoilState, useRecoilValue } from 'recoil';
 import { useComments } from '../recoil/comments';
-import useApi from '../services/api-interface-with-dashboard';
+import useApi from '../services/api';
 import { getData, useStorage } from '../services/dataManagement';
 import { capture } from '../services/sentry';
 import { useActions } from './actions';
