+++ dashboard/src/recoil/places.js
@@ -1,6 +1,6 @@
 import { atom, useRecoilState } from 'recoil';
 import { useRelsPerson } from '../recoil/relPersonPlace';
-import useApi from '../services/api-interface-with-dashboard';
+import useApi from '../services/api';
 import { getData, useStorage } from '../services/dataManagement';
 import { capture } from '../services/sentry';
