+++ dashboard/src/recoil/territoryObservations.js
@@ -1,5 +1,5 @@
 import { organisationState } from './auth';
-import useApi from '../services/api-interface-with-dashboard';
+import useApi from '../services/api';
 import { getData, useStorage } from '../services/dataManagement';
 import { capture } from '../services/sentry';
 import { atom, selector, useRecoilState, useRecoilValue } from 'recoil';
