+++ dashboard/src/scenes/team/view.js
@@ -13,7 +13,7 @@ import Box from '../../components/Box';
 import BackButton from '../../components/backButton';
 import NightSessionModale from '../../components/NightSessionModale';
 import { useAuth } from '../../recoil/auth';
-import useApi from '../../services/api-interface-with-dashboard';
+import useApi from '../../services/api';
 
 const View = () => {
   const [team, setTeam] = useState(null);
