+++ dashboard/src/scenes/user/list.js
@@ -16,7 +16,7 @@ import CreateWrapper from '../../components/createWrapper';
 import SelectCustom from '../../components/SelectCustom';
 import TagTeam from '../../components/TagTeam';
 import { useAuth } from '../../recoil/auth';
-import useApi from '../../services/api-interface-with-dashboard';
+import useApi from '../../services/api';
 
 const List = () => {
   const [users, setUsers] = useState(null);
